requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.12.1
babel==2.18.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
courlan==1.3.2
cryptography==46.0.7
dateparser==1.2.2
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.8
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h2==4.3.0
hpack==4.1.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
langchain-core==0.3.84
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.0.3
lxml_html_clean==0.4.4
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
openai==2.31.0
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
postgrest==2.28.3
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyiceberg==0.10.0
PyJWT==2.12.1
PyMuPDF==1.26.5
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-frontmatter==1.1.0
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
realtime==2.28.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.2
tiktoken==0.12.0
tld==0.13.2
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
